Fake jobs by environment recalibrated
Some info on screen

+++ src/main/java/br/com/gbvbahia/threads/monitor/persistence/repository/ProcessorRepository.java
@@ -20,5 +20,5 @@ public interface ProcessorRepository extends JpaRepository<Processor, Long> {
   
   List<Processor> findByProcessStatusAndUpdatedAtBefore(ProcessStatus processStatus, LocalDateTime updatedAt);
   
-  Long countByProcessStatus(ProcessStatus processStatus);
+  Integer countByProcessStatus(ProcessStatus processStatus);
 }
